Update retrieval location of rdf-toolkit.jar
The original download site is temporarily not providing a file needed to
implement CASE's normalization tests.

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

References:
* [ONT-383] Find alternative retrieval point for rdf-toolkit.jar

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>