Add UCO normalization confirmation to make check recursion
UCO has not yet formally adopted or encoded rdf-toolkit into its
workflow, but a proposal is under consideration for UCO 0.7.0.

Informal UCO incorporation of rdf-toolkit to date has already normalized
the content of the develop branch, so new content being merged in can be
tested for isolated normalization issues.  (This was confirmed by
running the CI tests of the CASE-Examples-QC repository and finding no
errors reported.)

References:
* [ONT-419] (CP-28) Adopt UCO 0.6.0
* [UCO OC-98] (CP-39) UCO should implement CI including at least
  rdf-toolkit runs

Signed-off-by: Alex Nelson