Validate kb.ttl
This patch adds a CASE validation check against the aggregate of all
examples between CASE-Examples, the CASE website, and CASE-Corpora.  A
script is added to perform a subset of RDFS entailment, found necessary
due to SHACL validation of subproperties requiring RDFS entailment to
preprocess a graph when reviewing subproperty hierarchies.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>