Remove normalization step
JSON normalization has been delegated to `pre-commit`.

This patch also bumps CASE-Examples' pointer to pick up a missed prefix
in one of the examples that caused a namespace conflict.

A follow-on patch will regenerate Make-managed files.

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>